Object library: array_schema
