+++ test/src/unit-capi-consolidation.cc
@@ -36,7 +36,7 @@
 #include "tiledb/sm/c_api/tiledb.h"
 #include "tiledb/sm/enums/encryption_type.h"
 #include "tiledb/sm/misc/time.h"
-#include "tiledb/sm/misc/utils.h"  // get_timestamp_range
+#include "tiledb/storage_format/uri/parse_uri.h"
 
 #include <climits>
 #include <cstring>
