+++ tiledb/CMakeLists.txt
@@ -242,6 +242,7 @@ set(TILEDB_CORE_SOURCES
   ${TILEDB_CORE_INCLUDE_DIR}/tiledb/sm/tile/generic_tile_io.cc
   ${TILEDB_CORE_INCLUDE_DIR}/tiledb/sm/tile/tile_metadata_generator.cc
   ${TILEDB_CORE_INCLUDE_DIR}/tiledb/sm/tile/writer_tile.cc
+  ${TILEDB_CORE_INCLUDE_DIR}/tiledb/storage_format/uri/parse_uri.cc
 )
 list(APPEND TILEDB_CORE_SOURCES ${TILEDB_COMMON_SOURCES})
 
@@ -315,6 +316,7 @@ set_source_files_properties(
   ${TileDB_blosc_SOURCES}
   PROPERTIES COMPILE_OPTIONS "${TileDB_blosc_COMPILE_OPTIONS}")
 
+add_subdirectory(storage_format)
 
 ############################################################
 # Build core objects as a reusable object library
@@ -513,7 +515,6 @@ if (SANITIZER)
       -fsanitize=${SANITIZER}
   )
 endif()
-
 
 # Coverage linker flags
 if (CMAKE_BUILD_TYPE MATCHES "Coverage")
