+++ tiledb/sm/array/array_directory.cc
@@ -35,8 +35,8 @@
 #include "tiledb/common/stdx_string.h"
 #include "tiledb/sm/filesystem/vfs.h"
 #include "tiledb/sm/misc/parallel_functions.h"
-#include "tiledb/sm/misc/utils.h"
 #include "tiledb/sm/misc/uuid.h"
+#include "tiledb/storage_format/uri/parse_uri.h"
 
 using namespace tiledb::common;
