+++ tiledb/sm/array_schema/array_schema.cc
@@ -47,7 +47,7 @@
 #include "tiledb/sm/filter/compression_filter.h"
 #include "tiledb/sm/misc/hilbert.h"
 #include "tiledb/sm/misc/time.h"
-#include "tiledb/sm/misc/utils.h"  // get_timestamp_range
+#include "tiledb/storage_format/uri/parse_uri.h"
 
 #include <cassert>
 #include <iostream>
