+++ tiledb/sm/consolidator/fragment_meta_consolidator.cc
@@ -40,6 +40,7 @@
 #include "tiledb/sm/storage_manager/storage_manager.h"
 #include "tiledb/sm/tile/generic_tile_io.h"
 #include "tiledb/sm/tile/tile.h"
+#include "tiledb/storage_format/uri/parse_uri.h"
 
 using namespace tiledb::common;
