+++ tiledb/sm/fragment/fragment_info.cc
@@ -41,6 +41,7 @@
 #include "tiledb/sm/misc/parallel_functions.h"
 #include "tiledb/sm/misc/time.h"
 #include "tiledb/sm/misc/utils.h"
+#include "tiledb/storage_format/uri/parse_uri.h"
 
 using namespace tiledb::sm;
 using namespace tiledb::common;
