+++ tiledb/sm/group/group_directory.cc
@@ -37,6 +37,7 @@
 #include "tiledb/sm/misc/parallel_functions.h"
 #include "tiledb/sm/misc/utils.h"
 #include "tiledb/sm/misc/uuid.h"
+#include "tiledb/storage_format/uri/parse_uri.h"
 
 using namespace tiledb::common;
