+++ tiledb/sm/misc/utils.h
@@ -69,53 +69,6 @@ std::string find_ca_certs_linux(const Posix& posix);
 }  // namespace https
 #endif
 
-namespace parse {
-
-/* ********************************* */
-/*          PARSING FUNCTIONS        */
-/* ********************************* */
-
-/**
- * Retrieves the timestamp range from the input
- * URI. For format version <= 2, only the range start is valid
- * (the range end is ignored).
- */
-Status get_timestamp_range(
-    const URI& uri, std::pair<uint64_t, uint64_t>* timestamp_range);
-
-/**
- * Retrieves the fragment name version.
- *  - Version 1 corresponds to format versions 1 and 2
- *      * __uuid_<t1>{_t2}
- *  - Version 2 corresponds to version 3 and 4
- *      * __t1_t2_uuid
- *  - Version 3 corresponds to version 5 or higher
- *      * __t1_t2_uuid_version
- */
-Status get_fragment_name_version(const std::string& name, uint32_t* version);
-
-/**
- * Retrieves the fragment version. This will work only for
- * name versions > 2, otherwise the function sets `version`
- * to UINT32_MAX.
- */
-Status get_fragment_version(const std::string& name, uint32_t* version);
-
-/** Returns `true` if the input string is a (potentially signed) integer. */
-bool is_int(const std::string& str);
-
-/** Returns `true` if the input string is an unsigned integer. */
-bool is_uint(const std::string& str);
-
-/** Converts the input value to string. */
-template <class T>
-std::string to_str(const T& value);
-
-/** Converts the input value of input type to string. */
-std::string to_str(const void* value, Datatype type);
-
-}  // namespace parse
-
 /* ********************************* */
 /*          TYPE FUNCTIONS           */
 /* ********************************* */
